Move PhotoResult.kt to datacollection/tasks/photo
