+++ ground/src/main/java/com/google/android/ground/ui/datacollection/tasks/photo/PhotoResult.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.android.ground.ui.editsubmission
+package com.google.android.ground.ui.datacollection.tasks.photo
 
 import android.graphics.Bitmap
